requirements.txt
pandas~=2.3.1
numpy~=2.3.2
deltalake~=1.1.4
pyarrow~=21.0.0
pyspark~=3.5.6
pyspark-stubs~=3.0.0.post3
delta-spark~=3.3.2
apache-airflow~=3.0.4
